\name{BSOINN-package}
\alias{BSOINN-package}
\docType{package}
\title{
\packageTitle{BSOINN}
}
\description{
This is a package for conducting Bayesian inference on the scalar on image regression with non-ignorable non-response
}
\details{
\tabular{ll}{
Package: \tab BSOINN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-11-29\cr
License: \tab GPL-2\cr
}

Function \code{BSOIFull} conducts Bayesian inference for a scalar on image regression with fully observed response.

Function \code{BSOIIN} conducts Bayesian inference for a scalar on image regression with ignorable non-response.

Function \code{BSOINN} conducts Bayesian inference for a scalar on image regression with non-ignorable non-response.

}
\author{
Feng X.N., Li T.F., Song X.Y., and Zhu H.T. (2017) Bayesian Scalar on Image Regression with Non-ignorable Non-response.
}

\keyword{ Bayesian approach, imaging data, instrumental variable, Markov chain Monte
Carlo, non-ignorable non-response}

